requirements.txt
Flask
Pillow
request
wechat-sdk
